cogl: Don't expose longs in the API (1/2)
unsigned long have quite a bad property in that they have different
sizes on 64 bits platforms depending on if you are on Linux (LP64) or on
Windows (LLP64).

This makes writing bindings for the CLR really hard as it only has fixed
types and no types which size would depend on the platform.

As the longs exposed in Cogl don't seem to need to be 64 bits
on 64 bits platforms, let's just use ints instead. Well, actually, even
if you wanted to use the 64 bits of longs and not care about Windows,
you couldn't because the API also has to work on Linux 32 bits.

Let's start by changing the type of the variable storing the number of
instances alive (per object type) to an int.

+++ cogl/cogl-object-private.h
@@ -117,7 +117,7 @@ struct _CoglObject
 #define COGL_OBJECT_COMMON_DEFINE_WITH_CODE(TypeName, type_name, code)  \
                                                                         \
 CoglObjectClass _cogl_##type_name##_class;                              \
-static unsigned long _cogl_object_##type_name##_count;                  \
+static unsigned int _cogl_object_##type_name##_count;                   \
                                                                         \
 static inline void                                                      \
 _cogl_object_##type_name##_inc (void)                                   \
